# Add option to search file in the file's target directory