--- a/lib/fastlane/plugin/forsis/helper/forsis_helper.rb
+++ b/lib/fastlane/plugin/forsis/helper/forsis_helper.rb
@@ -48,8 +48,8 @@ module Fastlane
           sonarqube_file.close
         end
 
-        def self.get_test_file_path(file_name, file_target)
-          `find #{file_target} -iname "#{file_name}.swift"`.gsub(/\n/, '')
+        def self.get_test_file_path(file_name, directory)
+          `find #{directory} -iname "#{file_name}.swift"`.gsub(/\n/, '')
         end
       end