Hello 👋,

We are using your plugin since few years and it worked quite well. 

## Problem
Our code coverage is growing and we noticed that the forsis lane is taking more and more time. 

After my investigation I found out that it was due to the `find` command that is used to recover the test file path linked to the test class. If you have many test files, like us, this command becomes very costly.

To give you an example, with more or less 1000 unit tests in 250 files, it takes 360 seconds to generate the file. 
In our case, this process is triggered every time we push to the repository on the CI which overload our CI. 

## Solution
In order to reduce this process duration I propose that we add a new option called `search_in_target_folder`.
By default the option will be set to `false` and it's optional so it will ensure the retro-compatibility of the plugin. 

With the default implementation the plugin will still search the file using the command `find .  -iname ...`. 

If the `search_in_target_folder` parameter is set to `true` it will use the `junit` report `testcase.name` property that looks like: 
```
<testsuite name='TestTarget.myTests' tests='2' failures='0'>
```
By splitting the name property we are able to retrieve the test target (most of the time associated to a directory) and specify to the `find` command in which directory to look in. 

By doing so we were able to drastically improve the plugin performance.
| Before | After |
|-------|-------|
| <img width="359" alt="Screenshot 2024-05-27 at 14 28 13" src="https://github.com/Tumiya/forsis/assets/10559183/37810ced-6add-4e15-9fb9-1bc7e08355c6"> | <img width="346" alt="Screenshot 2024-05-27 at 14 27 51" src="https://github.com/Tumiya/forsis/assets/10559183/62ffa03a-8e38-4a72-a7f5-c33deb82c83a"> |

Please, let me know if something is unclear.